Separate Form Control into separate component.
